fix: `EditorPatchProvider` patches are not saving their state when the "Activate" button is clicked.
When `PatchManagerSettingsProvider` calls `patch.EnablePatch()` the `forceSave` parameter defaults to `false` and that's passed around until it gets to `PatchManager.EnablePatch()` where short-circuit evaluation prevents `patchProviders[patchID].Instance.Persistent()` from being called. Replacing the `&&` with `||` fixes the issue, and the enable/disable state of a patch is now stored in the ProjectSettings asset.