feat: Customize the Project Settings window UI.
Each patch now features a slide toggle to enable/disable the patch, and the description, id, and type metadata was moved from a tooltip to an info box below the patch header.

Some Harmony assembly information now appears at the top of the window, and the footer layout was updated.

The location of the provider in the Project Settings window changed from "Needle/Editor Patching" to "Harmony".

The settings filename changed "EditorPatchingSettings" to "HarmonyPatchSettings".

BREAKING: This update has a dependency on the custom Imgui package, which provides the "slide toggle" control.